Rename samples module to sampler
