+++ doc/modules/sampler.rst
@@ -1,0 +1,6 @@
+.. currentmodule:: skopt.sampler
+
+.. _sampler:
+
+Sampling methods
+================
