--- doc/modules/samples.rst
@@ -1,6 +1,0 @@
-.. currentmodule:: skopt.samples
-
-.. _samples:
-
-Sampling methods
-================
