+++ examples/sampler/README.txt
@@ -1,0 +1,6 @@
+.. _sampler_examples:
+
+Initial sampling functions
+--------------------------
+
+Examples concerning the :mod:`skopt.sampler` module.
