+++ examples/sampler/initial-sampling-method-integer.py
@@ -1,12 +1,11 @@
 """
-==================================
-Comparing initial sampling methods
-==================================
+===================================================
+Comparing initial sampling methods on integer space
+===================================================
 
 Holger Nahrstaedt 2020 Sigurd Carlsen October 2019
 
 .. currentmodule:: skopt
-
 
 When doing baysian optimization we often want to reserve some of the
 early part of the optimization to pure exploration. By default the
@@ -19,19 +18,19 @@ take advantage of the fact that we know beforehand how many random
 points we want to sample. Then these points can be "spread out" in
 such a way that each dimension is explored.
 
-See also the example on an integer space
-:ref:`sphx_glr_auto_examples_initial_sampling_method_integer.py`
+See also the example on a real space
+:ref:`sphx_glr_auto_examples_initial_sampling_method.py`
 """
 
 print(__doc__)
 import numpy as np
-np.random.seed(123)
+np.random.seed(1234)
 import matplotlib.pyplot as plt
 from skopt.space import Space
-from skopt.samples import Sobol
-from skopt.samples import Lhs
-from skopt.samples import Halton
-from skopt.samples import Hammersly
+from skopt.sampler import Sobol
+from skopt.sampler import Lhs
+from skopt.sampler import Halton
+from skopt.sampler import Hammersly
 from scipy.spatial.distance import pdist
 
 #############################################################################
@@ -39,17 +38,18 @@ from scipy.spatial.distance import pdist
 def plot_searchspace(x, title):
     fig, ax = plt.subplots()
     plt.plot(np.array(x)[:, 0], np.array(x)[:, 1], 'bo', label='samples')
-    plt.plot(np.array(x)[:, 0], np.array(x)[:, 1], 'bo', markersize=80, alpha=0.5)
+    plt.plot(np.array(x)[:, 0], np.array(x)[:, 1], 'bs', markersize=40, alpha=0.5)
     # ax.legend(loc="best", numpoints=1)
     ax.set_xlabel("X1")
-    ax.set_xlim([-5, 10])
+    ax.set_xlim([0, 5])
     ax.set_ylabel("X2")
-    ax.set_ylim([0, 15])
+    ax.set_ylim([0, 5])
     plt.title(title)
+    ax.grid(True)
+
 
 n_samples = 10
-
-space = Space([(-5., 10.), (0., 15.)])
+space = Space([(0, 5), (0, 5)])
 space.set_transformer("normalize")
 
 #############################################################################
@@ -59,6 +59,7 @@ x = space.rvs(n_samples)
 plot_searchspace(x, "Random samples")
 pdist_data = []
 x_label = []
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("random")
 
@@ -69,26 +70,29 @@ x_label.append("random")
 sobol = Sobol()
 x = sobol.generate(space.dimensions, n_samples)
 plot_searchspace(x, 'Sobol')
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("sobol")
 
 #############################################################################
-# Classic Latin hypercube sampling
+# Classic latin hypercube sampling
 # --------------------------------
 
 lhs = Lhs(lhs_type="classic", criterion=None)
 x = lhs.generate(space.dimensions, n_samples)
 plot_searchspace(x, 'classic LHS')
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("lhs")
 
 #############################################################################
-# Centered Latin hypercube sampling
+# Centered latin hypercube sampling
 # ---------------------------------
 
 lhs = Lhs(lhs_type="centered", criterion=None)
 x = lhs.generate(space.dimensions, n_samples)
 plot_searchspace(x, 'centered LHS')
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("center")
 
@@ -99,6 +103,7 @@ x_label.append("center")
 lhs = Lhs(criterion="maximin", iterations=10000)
 x = lhs.generate(space.dimensions, n_samples)
 plot_searchspace(x, 'maximin LHS')
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("maximin")
 
@@ -109,6 +114,7 @@ x_label.append("maximin")
 lhs = Lhs(criterion="correlation", iterations=10000)
 x = lhs.generate(space.dimensions, n_samples)
 plot_searchspace(x, 'correlation LHS')
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("corr")
 
@@ -119,6 +125,7 @@ x_label.append("corr")
 lhs = Lhs(criterion="ratio", iterations=10000)
 x = lhs.generate(space.dimensions, n_samples)
 plot_searchspace(x, 'ratio LHS')
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("ratio")
 
@@ -129,6 +136,7 @@ x_label.append("ratio")
 halton = Halton()
 x = halton.generate(space.dimensions, n_samples)
 plot_searchspace(x, 'Halton')
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("halton")
 
@@ -139,6 +147,7 @@ x_label.append("halton")
 hammersly = Hammersly()
 x = hammersly.generate(space.dimensions, n_samples)
 plot_searchspace(x, 'Hammersly')
+print("empty fields: %d" % (36 - np.size(np.unique(x, axis=0), 0)))
 pdist_data.append(pdist(x).flatten())
 x_label.append("hammersly")
 
@@ -154,5 +163,5 @@ fig, ax = plt.subplots()
 ax.boxplot(pdist_data)
 plt.grid(True)
 plt.ylabel("pdist")
-_ = ax.set_ylim(0, 12)
+_ = ax.set_ylim(0, 6)
 _ = ax.set_xticklabels(x_label, rotation=45, fontsize=8)
