+++ examples/sampler/sampling_comparison.py
