--- examples/samples/README.txt
@@ -1,6 +1,0 @@
-.. _samples_examples:
-
-Initial sampling functions
---------------------------
-
-Examples concerning the :mod:`skopt.samples` module.
