+++ skopt/sampler/base.py
