+++ skopt/sampler/lhs.py
